requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
